Use newer support for Guava `CharMatcher`s since the constants were removed in Guava 26.0 and the preferred access pattern is the methods that were added in Guava 19.0
This makes the minimum version of the Cassandra Java Driver go up and supports Guava 19+.
Ideally, these methods should be updated to not require use of `CharMatcher` to not couple with the Guava version that the Cassandra driver uses.
Some hackery could be added to check if `CharMatcher.WHITESPACE` exists for Guava 18 and the current Cassandra version and then use `CharMatcher.whitespace()` on more recent versions, but this would increase complexity.