Attempt to read locks from cassandra table before inserting
If the insert locks fails, there's still a chance the table will contain
locks. This makes retrying the run difficult since the tables will have
an abandoned lock that requires manually cleaning up.

Attempt to read the locks at the desired consistency first. If it
succeeds, ignore the result and continue trying to insert locks with
existing success/failure logic. If the read locks fails, exit in a way
that has not altered tables and can be gracefully retried.

This will not handle all error cases, but helps in situations where (for
example) consistency cannot be achieved temporarily - when nodes are returned
to service the cqlmigrate can be reattempted automatically without
requiring manual lock cleanup.

See https://github.com/sky-uk/cqlmigrate/issues/87